requirements-dev.txt
# Dipendenze per sviluppo e testing
pytest==7.4.3
pytest-cov==4.1.0
pytest-flask==1.3.0

